extract RasterAssetInfo class into lib/stac.py
